\name{FUCHIKOMA-package}
\alias{FUCHIKOMA-package}
\alias{FUCHIKOMA}
\docType{package}
\title{
\packageTitle{FUCHIKOMA}
}
\description{
\packageDescription{FUCHIKOMA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FUCHIKOMA}
\packageIndices{FUCHIKOMA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FUCHIKOMA}

Maintainer: \packageMaintainer{FUCHIKOMA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
