\name{Marioni}
\alias{Marioni}
\docType{data}
\title{Marioni's dataset}
\description{
  This is a reduced version for the RNA-seq count data from Marioni et al. (2008) along with additional annotation such as gene biotype, gene length, GC content, chromosome, start position and end position for genes in chromosomes I to IV. The expression data consists of 10 samples from kidney and liver tissues. There are five technical replicates (lanes) per tissue.
}
\usage{
data(Marioni)
}
\references{
Marioni, J.C. and Mason, C.E. and Mane, S.M. and Stephens, M. and Gilad, Y. (2008) RNA-seq: an assessment of technical reproducibility and comparison with gene expression arrays. \emph{Genome Research}, \bold{18}: 1509--1517.
}
\keyword{datasets}