\name{FUCHIKOMA-package}
\alias{FUCHIKOMA-package}
\docType{package}
\title{
\packageTitle{FUCHIKOMA}
}
\description{
\packageDescription{FUCHIKOMA}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{FUCHIKOMA}
\packageIndices{FUCHIKOMA}
The main function is \code{\link{FUCHIKOMA}}, which returns an object containing the calculation results.
}
\author{
\packageAuthor{FUCHIKOMA}

Maintainer: \packageMaintainer{FUCHIKOMA}
}
\references{
Koki Tsuyuzaki et al. (2015) Fuchikoma: Detection of Differentially Expressed Genes in one of multiple clusters using BAHSIC and Diffusion Map. R package version 1.0.0

Laleh Haghverdi et al. (2015) Diffusion maps for high-dimensional single-cell analysis of differentiation data. Bioinformatics, 31(18), 2989-2998

Le Song et al. (2007) Gene selection via the BAHSIC family of algorithms, Bioinformatics, 23(13), i490-i498

Y-h Taguchi et al. (2015) Principal component analysis-based unsupervised feature extraction applied to in silico drug discovery for posttraumatic stress disorder-mediated heart disease, BMC Bioinformatics, 16(139)

Diego Adhemar Jaitin, Ephraim Kenigsberg, Hadas Keren-Shaul, Naama Elefant, Franziska Paul, Irina Zaretsky, Alexander Mildner, Nadav Cohen, Steffen Jung, Amos Tanay, Ido Amit (2014) Massively Parallel Single-Cell RNA-Seq for Marker-Free Decomposition of Tissues into Cell Types. \emph{Science}, \bold{343 (6172)}: 776-779

Arthur Gretton et al. (2007) A Kernel Statistical Test of Independence, NIPS 21

Aaditya Ramdas et al. (2015) Nonparametric Independence Testing for Small Sample Sizes, IJCAI-15
}
\keyword{ package }
\seealso{
\code{\link[destiny]{DiffusionMap}}
}