\docType{methods}
\name{simulateDropoutCounts}
\alias{simulateDropoutCounts}
\title{Generate simulation data from negative binomial (NB) distribution with Dropout events}

\description{
This function generates simulation data with arbitrary defined experimental condition.
}
\usage{
simulateDropoutCounts(Ngene = 10000, makeDEG=TRUE, PDEG = 0.2, DEG.assign = c(0.5, 0.5), DEG.foldchange = c(4, 4), replicates = c(3, 3), Lambda = 0.1)
}
\arguments{
  \item{Ngene}{No. of gene (defalult:10000)}
  \item{makeDEG}{Whether make DEG or not (defalult:TRUE)}
  \item{PDEG}{Percentage of DEG (defalult:0.2)}
  \item{DEG.assign}{Ratio of DEG in each cluster (defalult:c(0.5, 0.5))}
  \item{DEG.foldchange}{Fold change  in each cluster (defalult:c(4, 4))}
  \item{replicates}{No of repeated measurement in each cluster (defalult:c(3, 3))}
  \item{Lambda}{(defalult:0.1)}
 }

\value{
\item{DEG}{DEG list}
\item{FC}{Design matrix specifying fold change}
\item{Simcount}{Count data.frame generated by this function}
}

\author{
\packageAuthor{fuchikoma}

Maintainer: \packageMaintainer{fuchikoma}
}

\examples{
res1 <- simulateDropoutCounts(Ngene = 10000, replicates = c(3, 3), PDEG = 0.1, DEG.assign = c(0.5, 0.5), DEG.foldchange = c(4, 4), Lambda = 0.1)
str(res1)
}